feat: Add font-face declarations for Bengali and English languages
This commit adds font-face declarations for the Bengali and English languages in the index.scss file. The Bengali font-family is set to 'Bijoy' and the English font-family is set to 'Roboto'. This will ensure that the correct fonts are applied based on the language specified in the HTML elements.

The tailwind.config.js file is also updated to extend the theme with a new fontFamily property. The 'Roboto' and 'Bijoy' fonts are added to the existing 'sans' font-family, along with 'Arial' and 'sans-serif'.